backend/pyproject.toml
[project]
name = "parsewithai-backend"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "httpx>=0.27",
  "python-jose[cryptography]>=3.3",
  "asyncpg>=0.30",
  "sqlalchemy>=2.0",
  "langchain>=0.3",
  "langchain-google-genai>=2.0",
  "langchain-pinecone>=0.2",
  "langchain-community>=0.3",
  "langchain-text-splitters>=0.3",
  "pinecone>=5.0",
  "pypdf>=5.0",
  "python-multipart>=0.0.12",
  "python-magic-bin>=0.4.14; sys_platform == 'win32'",
  "python-magic>=0.4.27; sys_platform != 'win32'",
  "upstash-redis>=1.2",
  "tenacity>=9.0",
  "sse-starlette>=2.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "pytest-cov>=5.0",
  "respx>=0.21",
  "ruff>=0.7",
  "mypy>=1.13",
  "psycopg[binary]>=3.2",
  "reportlab>=4.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I","B","UP","SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
